created test client trait to remove duplicate code
